+++ tests/ttnn/unit_tests/gtests/accessor/test_accessor_benchmarks.cpp
@@ -12,7 +12,7 @@
 #include <cstdint>
 #include <tt-metalium/shape.hpp>
 #include <tt-metalium/distributed.hpp>
-#include "impl/profiler/tt_metal_profiler.hpp"
+#include "impl/profiler/tt_metal_profiler_internal.hpp"
 #include <tt-metalium/buffer_distribution_spec.hpp>
 
 #include <tt-metalium/tensor_accessor_args.hpp>
