+++ tt_metal/impl/profiler/tt_metal_profiler_internal.cpp
@@ -9,8 +9,7 @@
 #include <mesh_workload.hpp>
 #include <mesh_command_queue.hpp>
 #include <tt_metal.hpp>
-#include "impl/profiler/tt_metal_profiler.hpp"
-#include <algorithm>
+#include "impl/profiler/tt_metal_profiler_internal.hpp"
 #include <chrono>
 #include <cmath>
 #include <cstdint>
